@@ -16,13 +16,13 @@ public class Main {
 
     public static void createUserAndSave() {
         UserService userService = new UserServiceImpl();
-        User user = new User(1L, "Chlen2", "Lisi2i", (byte) 5);
+        User user = new User(1L, "Chlen2", "Lisi2i", (byte) 3);
         userService.saveUser(user.getName(), user.getLastName(), user.getAge());
-        User user2 = new User(1L, "Chlen3", "Lis3ii", (byte) 35);
+        User user2 = new User(1L, "Chlen3", "Lis3ii", (byte) 2);
         userService.saveUser(user2.getName(), user2.getLastName(), user2.getAge());
-        User user3 = new User(1L, "4", "Lisii3", (byte) 52);
+        User user3 = new User(1L, "4", "Lisii3", (byte) 3);
         userService.saveUser(user3.getName(), user3.getLastName(), user3.getAge());
-        User user4 = new User(1L, "Chlen5", "Lisi3i", (byte) 54);
+        User user4 = new User(1L, "Chlen5", "Lisi3i", (byte) 23);
         userService.saveUser(user4.getName(), user4.getLastName(), user4.getAge());
     }
 }
