@@ -1,15 +1,17 @@
 package jm.task.core.jdbc.service;
 
 import jm.task.core.jdbc.dao.UserDao;
+import jm.task.core.jdbc.dao.UserDaoHibernateImpl;
 import jm.task.core.jdbc.dao.UserDaoJDBCImpl;
 import jm.task.core.jdbc.model.User;
 
 import java.util.List;
 
 public class UserServiceImpl implements UserService {
-  private  final UserDao userDao = new UserDaoJDBCImpl();
+    private final UserDao userDao = new UserDaoHibernateImpl();
+
     public void createUsersTable() {
-    userDao.createUsersTable();
+        userDao.createUsersTable();
     }
 
     public void dropUsersTable() {
@@ -17,7 +19,7 @@ public class UserServiceImpl implements UserService {
     }
 
     public void saveUser(String name, String lastName, byte age) {
-        userDao.saveUser(name,lastName,age);
+        userDao.saveUser(name, lastName, age);
     }
 
     public void removeUserById(long id) {
@@ -27,6 +29,7 @@ public class UserServiceImpl implements UserService {
     public List<User> getAllUsers() {
         return userDao.getAllUsers();
     }
+
     public void cleanUsersTable() {
         userDao.cleanUsersTable();
     }
